backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0,<0.120.0
uvicorn[standard]>=0.27.0,<0.30.0
python-multipart>=0.0.6,<1.0.0

# AI APIs
anthropic>=0.21.0,<1.0.0
openai>=1.12.0,<2.0.0

# HTTP client
httpx>=0.26.0,<1.0.0

# Data validation
pydantic>=2.5.0,<3.0.0
email-validator>=2.1.0,<3.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0

# Web scraping
beautifulsoup4>=4.12.0,<5.0.0
praw>=7.7.0,<8.0.0
tweepy>=4.14.0,<5.0.0
feedparser>=6.0.0,<7.0.0

# Database
sqlalchemy>=2.0.25,<3.0.0
alembic>=1.13.0,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0

# Background tasks
celery[redis]>=5.3.0,<6.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Rate limiting
slowapi>=0.1.9,<1.0.0

# Payments
stripe>=7.0.0,<10.0.0

# YouTube integration
google-api-python-client>=2.100.0,<3.0.0
youtube-transcript-api>=0.6.0,<2.0.0
yt-dlp>=2024.1.0

# Production server (optional but recommended)
gunicorn>=21.0.0,<23.0.0
